Moved CSV methods to utils
